Added more code for goalieUpdateDB and its test

+++ fantasy-hockey-hidden-gems-java/main/ca/ualberta/emar/Goalie.java
@@ -219,7 +219,7 @@ public class Goalie {
 		this.timeOnIceQueue = timeOnIce;
 	}
 	
-	public float getTotalTimeOnIce(){
+	public int getTotalTimeOnIce(){
 		return totalTimeOnIce;
 	}
